äll--all

+++ user/unfollow.php
@@ -5,7 +5,7 @@
 
 	<title></title>
 </head>
-<body style="overflow-x: hidden;visibility: hidden;">
+<body style="overflow-x: hidden;">
 
 <?php
  session_start();
@@ -17,7 +17,9 @@ $sql1 = "SELECT `id`, `user_id`, `name`, `email`, `description`, `media`, `media
 $a = mysqli_query($conn, $sql1);
 $row = mysqli_fetch_assoc($a);
 
-
+$sql11 = "SELECT `id`, `user_id`, `name`, `email`, `gender`, `bio`, `dob`, `address`, `img`, `back_cover` FROM `profile` WHERE user_id = '".$_SESSION['id']."'";
+$a1 = mysqli_query($conn, $sql11);
+$row1 = mysqli_fetch_assoc($a1);
 
 
 $like = "INSERT INTO `notify`(`user_id`, `creator_id`, `title`, `icon`, `text`) VALUES ('".$_SESSION['id']."' , '".$row['user_id']."' , 'Unfollowing You' , '".$row1['img']."' , '@".$row1['name']."')";
@@ -27,13 +29,15 @@ $data = mysqli_fetch_assoc($run);
 
 
 
-$like1 = "DELETE FROM `follow` WHERE user_id = '".$_SESSION['id']."'";
+$like1 = "DELETE FROM `follow` WHERE user_id = '".$_SESSION['id']."' AND creator_id='".$_GET['creator_id']."'";
 $run1 = mysqli_query($conn, $like1);
 $data1 = mysqli_fetch_assoc($run1);
 
+echo $like1;
+
 echo '<script type="text/javascript">';
 
-echo 'window.location.href = "'. "http://205.147.99.187/FriendsBook/user/veiw.php?creator_id=".$_GET['creator_id']."".'";';
+// echo 'window.location.href = "home.php?"';
 
 echo '</script>';
